# Slightly optimized Url.Clean