Unfortunately the code slows down by a couple percent if I use
temp variables ... for some reason

name  old time/op    new time/op    delta
C-8     27.9ns ±11%    22.6ns ± 8%  -19.12%  (p=0.000 n=10+9)

name  old alloc/op   new alloc/op   delta
C-8      32.0B ± 0%     32.0B ± 0%     ~     (all equal)

name  old allocs/op  new allocs/op  delta
C-8       1.00 ± 0%      1.00 ± 0%     ~     (all equal)