@MStoykov 2020-07-27
I changed the `C` from this old [benchmark](https://gist.github.com/MStoykov/280f527fdf0834d275d40b5133064e4b) to 
```golang
func C(u *url.URL, out string) string {
        if u == nil || u.User == nil {
                return out
        }

        if password, passwordOk := u.User.Password(); passwordOk {
                return out[:len(u.Scheme)+3] + "****:****" + out[len(u.Scheme)+4+len(u.User.Username())+len(password):]
        }

        return out[:len(u.Scheme)+3] + "****" + out[len(u.Scheme)+3+len(u.User.Username()):]
}
```

Unfortunately, as I mentioned, having temp variables seems to reduce the improvement by around 1-5% ... so I dropped them as I think it's somewhat obvious what we are doing, and this is a kind of a hot path